AUtomated REsume Checker/requirements.txt
streamlit
pandas
numpy
spacy
sentence-transformers
transformers
faiss-cpu
pdfplumber
python-docx
torch
scikit-learn
python-dotenv
pyyaml
sqlalchemy
bcrypt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
pydantic[email]
